% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mojxml_download.R
\name{mojxml_download}
\alias{mojxml_download}
\title{Download MOJ XML}
\usage{
mojxml_download(x, exdir, progress = TRUE, ...)
}
\arguments{
\item{x}{File names of MOJ XML or a data frame created with \code{mojxml_find()}.}

\item{exdir}{A directory name to download data. If it does not exist, it will
be created.}

\item{progress}{Whether to show a progress bar. Passed to \code{\link[purrr:map]{purrr::walk()}}.}

\item{...}{Unused, for extensibility.}
}
\value{
The names of downloaded files invisibly.
}
\description{
Download MOJ XML
}
