% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mojxml_find.R
\name{mojxml_find}
\alias{mojxml_find}
\title{Find MOJ XML via 'kuwanauchi'}
\usage{
mojxml_find(city_code, lang_col_names = c("en", "ja"))
}
\arguments{
\item{city_code}{A city code in Japan.}

\item{lang_col_names}{Language of column names, English (\code{"en"}; default) or
Japanese (\code{"ja"}).}
}
\value{
A \code{mojxml_find} object, which is a data frame.
}
\description{
Find MOJ XML via 'kuwanauchi'
}
