% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mojxmlr-package.R
\docType{package}
\name{mojxmlr-package}
\alias{mojxmlr}
\alias{mojxmlr-package}
\title{mojxmlr: Download and Read MOJ XML}
\description{
You can download MOJ XML via 'kuwanauchi' and read them as a 'sf' object.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UchidaMizuki/mojxmlr}
  \item \url{https://uchidamizuki.github.io/mojxmlr/}
  \item Report bugs at \url{https://github.com/UchidaMizuki/mojxmlr/issues}
}

}
\author{
\strong{Maintainer}: Mizuki Uchida \email{uchidamizuki@vivaldi.net}

}
\keyword{internal}
