% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mojxml.R
\name{read_mojxml}
\alias{read_mojxml}
\title{Read MOJ XML}
\usage{
read_mojxml(
  file,
  type = c("fude_polygon", "fude_point", "fude_line", "admin_line", "map_info"),
  lang_col_names = c("en", "ja"),
  add_columns = TRUE
)
}
\arguments{
\item{file}{A file name of MOJ XML.}

\item{type}{The data type to be read, \code{"fude_polygon"} (default),
\code{"fude_polygon"}, \code{"fude_line"}, \code{"admin_line"}, or \code{"map_info"}.}

\item{lang_col_names}{Language of column names, English (\code{"en"}; default) or
Japanese (\code{"ja"}).}

\item{add_columns}{Whether to add MOJ XML attributes to columns? By default,
\code{TRUE}. If \code{FALSE}, they are added to \code{attributes}.}
}
\value{
A sf object, except when \code{type == "map_info"}. If
\code{type == "map_info"}, a tibble object.
}
\description{
Read MOJ XML
}
