% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mojxmlr-package.R
\docType{package}
\name{mojxmlr-package}
\alias{mojxmlr}
\alias{mojxmlr-package}
\title{mojxmlr: Download and Read MOJ XML}
\description{
You can download MOJ XML via 'kuwanauchi' and read them as a 'sf' object.
}
\author{
\strong{Maintainer}: Mizuki Uchida \email{uchidamizuki@vivaldi.net}

}
\keyword{internal}
